refactor: 숫자 구분자 서비스 추상화, 정수 숫자 구분자 클래스를 사용하도록 변경
- 숫자는 추상적인 개념임.
- 정수 숫자 구분자 클래스로의 변경을 통해 유연성 증가.
